Creating associate learning
This creates associate learning for link modules. So, if the admin creates a course and adds a face-to-face module. And then adds a link module to this course and sets it as associate learning. Then the user should not be able to start the link module unless the admin confirms the booking and the payment of the user.